refactor Message and Worker into the control package to collect usage classes into one package
