+++ src/main/java/gov/cida/cdat/control/Control.java
@@ -1,6 +1,5 @@
 package gov.cida.cdat.control;
 
-import gov.cida.cdat.message.Message;
 
 
 /**
