+++ src/main/java/gov/cida/cdat/service/Session.java
@@ -2,12 +2,12 @@ package gov.cida.cdat.service;
 
 import static akka.actor.SupervisorStrategy.stop;
 import gov.cida.cdat.control.Control;
+import gov.cida.cdat.control.Message;
 import gov.cida.cdat.control.SCManager;
 import gov.cida.cdat.control.Status;
 import gov.cida.cdat.control.Time;
 import gov.cida.cdat.exception.CdatException;
 import gov.cida.cdat.message.AddWorkerMessage;
-import gov.cida.cdat.message.Message;
 
 import java.lang.ref.WeakReference;
 import java.util.HashMap;
