+++ src/test/java/gov/cida/cdat/control/SCManagerControlFailTest.java
@@ -8,9 +8,7 @@ import gov.cida.cdat.control.SCManager;
 import gov.cida.cdat.control.Status;
 import gov.cida.cdat.io.container.DataPipe;
 import gov.cida.cdat.io.container.SimpleStreamContainer;
-import gov.cida.cdat.message.Message;
 import gov.cida.cdat.service.PipeWorker;
-import gov.cida.cdat.service.Worker;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
