+++ src/test/java/gov/cida/cdat/service/DelegatorAutoStartTest.java
@@ -4,8 +4,8 @@ package gov.cida.cdat.service;
 import gov.cida.cdat.TestUtils;
 import gov.cida.cdat.control.Control;
 import gov.cida.cdat.control.SCManager;
+import gov.cida.cdat.control.Worker;
 import gov.cida.cdat.exception.CdatException;
-import gov.cida.cdat.service.Worker;
 import static org.junit.Assert.*;
 
 import org.junit.Test;
