+++ src/test/java/gov/cida/cdat/service/DelegatorStatusTests.java
@@ -4,12 +4,12 @@ package gov.cida.cdat.service;
 import gov.cida.cdat.TestUtils;
 import gov.cida.cdat.control.Callback;
 import gov.cida.cdat.control.Control;
+import gov.cida.cdat.control.Message;
 import gov.cida.cdat.control.SCManager;
 import gov.cida.cdat.control.Status;
+import gov.cida.cdat.control.Worker;
 import gov.cida.cdat.exception.CdatException;
-import gov.cida.cdat.message.Message;
 import gov.cida.cdat.service.PipeWorker;
-import gov.cida.cdat.service.Worker;
 import static org.junit.Assert.*;
 
 import org.junit.Test;
