+++ src/test/java/gov/cida/cdat/service/SessionWorkerHistoryTests.java
@@ -4,10 +4,11 @@ import static org.junit.Assert.*;
 import gov.cida.cdat.TestUtils;
 import gov.cida.cdat.control.Callback;
 import gov.cida.cdat.control.Control;
+import gov.cida.cdat.control.Message;
 import gov.cida.cdat.control.SCManager;
 import gov.cida.cdat.control.Status;
+import gov.cida.cdat.control.Worker;
 import gov.cida.cdat.exception.CdatException;
-import gov.cida.cdat.message.Message;
 
 import org.junit.Before;
 import org.junit.Test;
